renaming 'Quiz' to 'QuizLegacy' and 'QuizAlt' to 'Quiz'
