+++ resources/assets/components/LegacyQuiz/Conclusion.js
