+++ resources/assets/components/LegacyQuiz/helpers.js
