+++ resources/assets/components/LegacyQuiz/index.js
@@ -1,0 +1,7 @@
+// The <LegacyQuizContainer> can match a route to a particular
+// quiz from the campaign feed in the Redux store.
+export LegacyQuizContainer from './LegacyQuizContainer';
+
+// The <LegacyQuiz> component accepts a JSON blob from the
+// Contentful Quiz content type & renders the appropriate content.
+export default from './LegacyQuiz';
