+++ resources/assets/components/Page/CampaignPage/CampaignPage.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { Switch, Route, Redirect } from 'react-router-dom';
 
 import { FeedContainer } from '../../Feed'; // @TODO: rename to ActivityFeed or ActivityPage...
-import { QuizContainer } from '../../Quiz';
+import { LegacyQuizContainer } from '../../LegacyQuiz';
 import BlockPageContainer from '../BlockPage';
 import { isCampaignClosed } from '../../../helpers';
 import { ActionPageContainer } from '../ActionPage';
@@ -64,7 +64,7 @@ const CampaignPage = (props) => {
           />
           <Route path={`${match.url}/pages/:slug`} component={CampaignSubPageContainer} />
           <Route path={`${match.url}/blocks/:id`} component={BlockPageContainer} />
-          <Route path={`${match.url}/quiz/:slug`} component={QuizContainer} />
+          <Route path={`${match.url}/quiz/:slug`} component={LegacyQuizContainer} />
           <Route
             path={`${match.url}/modal/:id`}
             render={(routingProps) => {
