+++ resources/assets/components/Quiz/QuizConclusion.js
