--- resources/assets/components/Quiz/index.js
@@ -1,7 +1,0 @@
-// The <QuizContainer> can match a route to a particular
-// quiz from the campaign feed in the Redux store.
-export QuizContainer from './QuizContainer';
-
-// The <Quiz> component accepts a JSON blob from the
-// Contentful Quiz content type & renders the appropriate content.
-export default from './Quiz';
