+++ resources/assets/components/Quiz/quiz.scss
@@ -3,11 +3,7 @@
 .quiz {
   margin: 0 $half-spacing;
 
-  .quiz__introduction {
-    text-align: center;
-  }
-
-  .quiz__subtitle {
+  .quiz__heading {
     color: $primary-color;
     font-size: $font-large;
     font-weight: $weight-bold;
@@ -19,10 +15,6 @@
     font-size: $font-hero;
     text-transform: uppercase;
     margin-bottom: $base-spacing;
-  }
-
-  .quiz__description {
-    margin-bottom: $base-spacing * 2;
   }
 
   .question {
@@ -48,17 +40,20 @@
     flex-wrap: wrap;
   }
 
-  .answer {
+  .choice {
     margin-bottom: $half-spacing;
     font-weight: $weight-bold;
-    text-align: center;
     text-decoration: none;
     width: 100%;
     outline: 0;
 
-    @include media($medium) {
+    &.background-image {
       flex: 0 0 33%;
-      padding: 0 $half-spacing $half-spacing $half-spacing;
+      padding: 0 $half-spacing $half-spacing 0;
+
+      @include media($small) {
+        flex: 0 0 50%;
+      }
     }
 
     &:hover {
@@ -71,7 +66,7 @@
     }
   }
 
-  .conclusion {
+  .quiz-conclusion {
     margin-bottom: $base-spacing;
 
     .conclusion__item {
@@ -79,18 +74,23 @@
       vertical-align: middle;
       box-sizing: border-box;
 
+      @include media($small) {
+        width: 100%;
+      }
+
       @include media($tablet) {
         &.-one-third {
-          width: 33%;
+          width: 50%;
         }
 
         &.-two-thirds {
-          width: 66%;
+          width: 50%;
         }
       }
 
       button {
         width: 100%;
+        min-height: 65px;
       }
     }
   }
